Refactor test imports and enhance user authentication tests
- Updated import statements in test files to use the new utility functions from `tests.utils`.
- Added a second user registration and login process in the profiles test to improve authentication coverage.
- This refactor aims to streamline test organization and enhance the robustness of user-related tests.
