+++ src/profiles/profiles.test.ts
@@ -2,6 +2,7 @@ import { afterAll, beforeAll, describe, expect, it } from "bun:test";
 import { treaty } from "@elysiajs/eden";
 import { app } from "@/core/app";
 import { db } from "@/core/db";
+import { expectSuccess, expectToBeDefined } from "@/shared/utils/tests.utils";
 
 // Create type-safe API client with Eden Treaty
 const { api } = treaty(app);
@@ -20,6 +21,7 @@ const testUser2 = {
 };
 
 let authToken: string;
+let authToken2: string;
 
 beforeAll(async () => {
 	// Reset database
@@ -29,11 +31,21 @@ beforeAll(async () => {
 	const reg1 = await api.users.post({ user: testUser });
 	authToken = reg1.data?.user?.token ?? "";
 
+	// Register second user
+	const reg2 = await api.users.post({ user: testUser2 });
+	authToken2 = reg2.data?.user?.token ?? "";
+
 	// Login first user (to ensure token is valid)
 	const login1 = await api.users.login.post({
 		user: { email: testUser.email, password: testUser.password },
 	});
 	authToken = login1.data?.user?.token ?? "";
+
+	// Login second user (to ensure token is valid)
+	const login2 = await api.users.login.post({
+		user: { email: testUser2.email, password: testUser2.password },
+	});
+	authToken2 = login2.data?.user?.token ?? "";
 });
 
 describe("Profile Tests", () => {
